Add new device endpoint still wip

+++ pos-monitoring/src/main/resources/db/migration/V2__Create_Pos_Device_Status_Change_Log_Table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE pos_device_status_change_log (
-    id BIGINT NOT NULL PRIMARY KEY,
+    id SERIAL NOT NULL PRIMARY KEY,
     pos_device_id INT NOT NULL,
     online BOOLEAN NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
